move ParameterAccessFlags to representation pkg
